Initial port architecture for Pouch
